rm engine-primitives dep
