+++ crates/engine/primitives/src/lib.rs
@@ -13,7 +13,6 @@ extern crate alloc;
 
 use alloy_consensus::BlockHeader;
 use reth_errors::ConsensusError;
-use reth_evm::{execute::ExecutableTxFor, ConfigureEvm, EvmEnvFor, ExecutionCtxFor};
 use reth_payload_primitives::{
     EngineApiMessageVersion, EngineObjectValidationError, InvalidPayloadAttributesError,
     NewPayloadError, PayloadAttributes, PayloadOrAttributes, PayloadTypes,
@@ -23,6 +22,7 @@ use reth_trie_common::HashedPostState;
 use serde::{de::DeserializeOwned, Serialize};
 
 // Re-export [`ExecutionPayload`] moved to `reth_payload_primitives`
+pub use reth_evm::{ConfigureEngineEvm, ExecutableTxIterator};
 pub use reth_payload_primitives::ExecutionPayload;
 
 mod error;
@@ -171,36 +171,4 @@ pub trait PayloadValidator<Types: PayloadTypes>: Send + Sync + Unpin + 'static {
         }
         Ok(())
     }
-}
-
-/// [`ConfigureEvm`] extension providing methods for executing payloads.
-pub trait ConfigureEngineEvm<ExecutionData>: ConfigureEvm {
-    /// Returns an [`EvmEnvFor`] for the given payload.
-    fn evm_env_for_payload(&self, payload: &ExecutionData) -> EvmEnvFor<Self>;
-
-    /// Returns an [`ExecutionCtxFor`] for the given payload.
-    fn context_for_payload<'a>(&self, payload: &'a ExecutionData) -> ExecutionCtxFor<'a, Self>;
-
-    /// Returns an [`ExecutableTxIterator`] for the given payload.
-    fn tx_iterator_for_payload(&self, payload: &ExecutionData) -> impl ExecutableTxIterator<Self>;
-}
-
-/// Iterator over executable transactions.
-pub trait ExecutableTxIterator<Evm: ConfigureEvm>:
-    Iterator<Item = Result<Self::Tx, Self::Error>> + Send + 'static
-{
-    /// The executable transaction type iterator yields.
-    type Tx: ExecutableTxFor<Evm> + Clone + Send + 'static;
-    /// Errors that may occur while recovering or decoding transactions.
-    type Error: core::error::Error + Send + Sync + 'static;
-}
-
-impl<Evm: ConfigureEvm, Tx, Err, T> ExecutableTxIterator<Evm> for T
-where
-    Tx: ExecutableTxFor<Evm> + Clone + Send + 'static,
-    Err: core::error::Error + Send + Sync + 'static,
-    T: Iterator<Item = Result<Tx, Err>> + Send + 'static,
-{
-    type Tx = Tx;
-    type Error = Err;
 }
