+++ crates/optimism/evm/Cargo.toml
@@ -12,7 +12,6 @@ workspace = true
 
 [dependencies]
 # Reth
-reth-engine-primitives.workspace = true
 reth-payload-primitives.workspace = true
 reth-chainspec.workspace = true
 reth-evm = { workspace = true, features = ["op"] }
@@ -76,7 +75,6 @@ std = [
     "reth-evm/std",
     "alloy-rpc-types-engine/std",
     "op-alloy-rpc-types-engine/std",
-    "reth-engine-primitives/std",
 ]
 portable = ["reth-revm/portable"]
 rpc = ["reth-rpc-eth-api"]
